email_system/api/requirements.txt
fastapi
uvicorn
prometheus-client
kafka-python
pydantic